*BugFix* Made a bug fix which only shows 404 isntead of internal error when nion admin acccesses /pastes URL
*Changes*
We don't like that "New Paste" is now part of the project menu. Its already enough tabs in our projects menu - eg. new issue is to be request not part or projects menu also.
So we removed the entry from the menu and put "Redmine alike" a "new Paste" link on the upper right of the "Pastes" view tab, which is also part of the project menu (again: one tab is enough per modul)